[#4403][refactor] Move match_moe, sharding, load_weights to modular inference optimizer (#124)
* refactor: move match_moe_pattern, sharding, load_weights to modular inf optimizer


* refactor: move match_moe, sharding, load_weights to modular inference optimizer


* style: fix typo; remove redundant comments


* polish: keep moe skip test; move sharding configs to avoid circular imports;


* remove redundant comments


* style: rename class name


* fix quant_graph inputs; add type hints for shared_config


---------
