+++ tensorrt_llm/_torch/auto_deploy/config/default.yaml
@@ -27,6 +27,8 @@ transforms:
     stage: pattern_matcher
   match_attention_layout:
     stage: pattern_matcher
+  match_moe_pattern:
+    stage: pattern_matcher
   match_rope_pattern:
     stage: pattern_matcher
   match_rope_layout:
@@ -43,3 +45,17 @@ transforms:
     stage: pattern_matcher
   quantize_moe:
     stage: pattern_matcher
+  # TODO: Infer sharding parameters (tp_size, row/column sharding) from the model config.
+  detect_column_row_shard:
+    stage: sharding
+    simple_shard_only: false
+  detect_ep_shard:
+    stage: sharding
+  detect_dp_bmm_shard:
+    stage: sharding
+  # TODO: (hg) need to ensure run_shape_prop after sharding.
+  sharding_transform_executor:
+    stage: sharding
+    run_shape_prop: true
+  load_weights:
+    stage: weight_load
