+++ tensorrt_llm/_torch/auto_deploy/transform/optimizer.py
@@ -2,13 +2,16 @@
 
 from typing import Optional
 
+import torch.distributed as dist
 import torch.nn as nn
 from torch.fx import Graph, GraphModule
 
+from ..distributed import common as dist_ad
 from ..models.factory import ModelFactory
 from ..shim.interface import CachedSequenceInterface
 from .interface import (
     InferenceOptimizerConfig,
+    SharedConfig,
     Stages,
     StrictInferenceOptimizerConfig,
     TransformConfig,
@@ -20,6 +23,11 @@ class InferenceOptimizer:
     def __init__(self, factory: ModelFactory, config: InferenceOptimizerConfig):
         self.factory = factory
         self.config = self._clean_config(config)
+        if not dist.is_initialized():
+            local_rank, world_size = 0, 1
+        else:
+            local_rank, world_size = dist_ad.get_rank_world_size()
+        self.shared_config = SharedConfig(local_rank=local_rank, world_size=world_size)
 
     def _clean_config(self, config: InferenceOptimizerConfig) -> StrictInferenceOptimizerConfig:
         """Get a typed checked ("strict") config with sorted keys according to stages."""
@@ -68,7 +76,7 @@ class InferenceOptimizer:
             # instantiate transform
             transform = TransformRegistry.get(t_name)(t_config)
             # run transform
-            gm = transform(gm, cm, self.factory)
+            gm = transform(gm, cm, self.factory, self.shared_config)
 
         ############################################################################################
         # RETURN OPTIMIZED GRAPH
