+++ tensorrt_llm/_torch/auto_deploy/transformations/library/__init__.py
@@ -5,7 +5,6 @@ from .fused_moe import *
 from .fusion import *
 from .kvcache import *
 from .rms_norm import *
-from .sharding import *
 
 try:
     from .visualization import visualize_namespace
