+++ tensorrt_llm/_torch/auto_deploy/transformations/transform.py
@@ -7,25 +7,17 @@ import torch.nn as nn
 
 from ..compile import compile_and_capture
 from ..custom_ops.attention_interface import AttentionRegistry
-from ..distributed import common as dist_ad
 from ..llm_args import AutoDeployConfig
 from ..models.factory import ModelFactory
 from ..shim.interface import CachedSequenceInterface
 from ..transform.optimizer import InferenceOptimizer as ModularInferenceOptimizer
 from ..utils.logger import ad_logger
-from ._graph import canonicalize_graph, lift_to_meta, move_to_device
 from .library import (
-    ShardingConfig,
-    detect_column_row_shard,
-    detect_dp_bmm_shard,
-    detect_ep_shard,
     fuse_allreduce_residual_rmsnorm,
     fuse_collectives,
     fuse_rmsnorm,
     insert_cached_attention,
-    match_moe_pattern,
     resize_kv_cache,
-    sharding_transform_executor,
     update_in_out_nodes,
 )
 
@@ -53,6 +45,7 @@ class InferenceOptimizer:
         # RUN MODULAR INFERENCE OPTIMIZER FOR ALREADY-MIGRATED TRANSFORMS
         ############################################################################################
         # TODO (hg): default values that are not representable in YAML.
+        # move to the optimizer
         if "match_attention_layout" in self.ad_config.transforms:
             self.ad_config.transforms[
                 "match_attention_layout"
@@ -63,56 +56,17 @@ class InferenceOptimizer:
             ).get_attention_layout()
 
         new_optimizer = ModularInferenceOptimizer(self.factory, self.ad_config.transforms)
+
+        # TODO (hg): similar to above.
+        if "load_weights" in new_optimizer.config:
+            new_optimizer.config[
+                "load_weights"
+            ].checkpoint_device = self.ad_config.checkpoint_device
+            new_optimizer.config["load_weights"].device = cm.device
+
         egm = new_optimizer(cm)
 
         # TODO (lucaslie): continue moving legacy transforms to the new optimizer
-
-        ############################################################################################
-        # RUN PATTERN MATCHER TRANSFORMATIONS TO STANDARDIZE GRAPH REPRESENTATION
-        ############################################################################################
-
-        # Match MoE pattern
-        # TODO:remove quantized linear handling inside this transformation
-        match_moe_pattern(egm)
-
-        ############################################################################################
-        # RUN TRANSFORMATIONS ON STANDARDIZED GRAPH REPRESENTATION
-        ############################################################################################
-
-        local_rank, world_size = dist_ad.get_rank_world_size()
-
-        # TODO: Infer sharding parameters (tp_size, row/column sharding) from the model config.
-        sharding_config = ShardingConfig()
-
-        # run TP sharding across ranks
-        detect_column_row_shard(
-            egm, local_rank, world_size, sharding_config, self.ad_config.simple_shard_only
-        )
-
-        # run EP sharding across ranks
-        detect_ep_shard(egm, local_rank, world_size, sharding_config)
-
-        # run BMM sharding across ranks
-        detect_dp_bmm_shard(egm, local_rank, world_size, sharding_config)
-
-        sharding_transform_executor(egm, sharding_config)
-
-        # let's run a shape propagation pass to update the graph with correct meta values for
-        # subsequent optimization passes. Lift state_dict to meta as shape propagation involves device check
-        with lift_to_meta(egm):
-            canonicalize_graph(egm, shape_prop=True)
-
-        ############################################################################################
-        # MOVE MODEL AND LOAD WEIGHTS
-        ############################################################################################
-
-        # load weights
-        self.factory.load_or_random_init(egm, device=self.ad_config.checkpoint_device or cm.device)
-
-        # move remaining parts to device
-        move_to_device(egm, cm.device)
-        cm.to(cm.device)
-
         ############################################################################################
         # RUN POST-LOAD FUSION AND OPTIMIZATIONS
         ############################################################################################
