+++ tests/unittest/_torch/auto_deploy/_utils_test/_graph_test_helpers.py
@@ -11,7 +11,7 @@ from torch.fx import GraphModule
 from tensorrt_llm._torch.auto_deploy.custom_ops.attention_interface import SequenceInfo
 from tensorrt_llm._torch.auto_deploy.export import torch_export_to_gm
 from tensorrt_llm._torch.auto_deploy.models.factory import ModelFactory
-from tensorrt_llm._torch.auto_deploy.transformations.library.sharding import ShardingTransformInfo
+from tensorrt_llm._torch.auto_deploy.transform.library.sharding import ShardingTransformInfo
 
 
 class FakeFactory(ModelFactory):
